hosts ids get to Short rather than int

+++ template_java/src/main/java/cs451/Host.java
@@ -7,13 +7,13 @@ public class Host {
 
     private static final String IP_START_REGEX = "/";
 
-    private int id;
+    private short id;
     private String ip;
     private int port = -1;
 
     public boolean populate(String idString, String ipString, String portString) {
         try {
-            id = Integer.parseInt(idString);
+            id = Short.parseShort(idString);
 
             String ipTest = InetAddress.getByName(ipString).toString();
             if (ipTest.startsWith(IP_START_REGEX)) {
@@ -41,7 +41,7 @@ public class Host {
         return true;
     }
 
-    public int getId() {
+    public short getId() {
         return id;
     }
